feat(nodenv): gets node version from service
The idea is to have a endpoint to get the latest node version that we
are using at platanus. The endpoint is http://node.platan.us/latest

The generator gets the version of the service and generates the
corresponding `.node-version`

If the service is down, if fallback to whatever node version the
generator it's being run on. This is done via the process.version
variable.